currency model finished

+++ app/api/models/users.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer
+from sqlalchemy import BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.database import Base
@@ -6,6 +6,6 @@ from app.db.database import Base
 
 class User(Base):
     __tablename__ = "user"
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True)
     username: Mapped[str] = mapped_column(nullable=False)
     password: Mapped[str] = mapped_column(nullable=False)
